+++ packages/viness-react/src/router/index.ts
@@ -2,7 +2,25 @@ export * from './router.module'
 export * from './route.factory'
 
 export * from './hooks'
-
 export * from './router-react'
 
-export { Navigate, Outlet, Link, NavLink, useNavigate } from 'react-router-dom'
+export {
+    useLocation,
+    useMatches,
+    useMatch,
+    useNavigate,
+    useParams,
+    useOutlet,
+    useSearchParams,
+    Link,
+    NavLink,
+    Outlet,
+    Route,
+    Routes,
+    Navigate,
+    Await,
+    ScrollRestoration,
+    redirect,
+    generatePath,
+    matchPath
+} from 'react-router-dom'